requirements.txt
pycryptodome
prettytable
